FindIgnOGRE2: prefer versioned component libraries
Both ogre1.9 and ogre2.1 install component libraries named
libOgreOverlay. For ogre1.9, an unversioned symbolic link is
installed directly into the lib folder, while in our debian
libogre-2.1-dev and brew ogre2.1 packages, an unversioned symbolic
link is installed in lib/OGRE-2.1.

This find module uses find_library to find Ogre component libraries
with HINTS pointing to the lib/OGRE-2.1 subfolder, which is helful
but limited because the CMAKE_PREFIX_PATH is searched before
the HINTS. When building ignition-rendering with homebrew, the
CMAKE_PREFIX_PATH may include /usr/local, which causes the
libOgreOverlay library from ogre1.9 to be found before the
version from ogre2.1.

On macOS, the following are in /usr/local/lib:
    /usr/local/lib/libOgreOverlay.1.9.0.dylib
    /usr/local/lib/libOgreOverlay.2.1.0.dylib
    /usr/local/lib/libOgreOverlay.dylib (sym-link to 1.9.0)

This causes find_library to prefer a library name with the version
appended before the suffix, while still accepting unversioned names
if a versioned name is not found.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>