Fix FindIgnOgre on Windows when not using vcpkg (#124)
The current version of FindIgnOgre contains a vcpkg-specific workaround on the location of the OgreMain
library, that is fixed to point to the "${OGRE_LIBRARY_DIRS}/manual-link/" directory, that it is its location when
Ogre is installed by vcpkg. However, this creates a problem when installing Ogre with other package managers
on Windows, for example when using conda-forge or conan. To mitigate this problem, we only set the location of
the OgreMain library to be "${OGRE_LIBRARY_DIRS}/manual-link/" only if this directory exists, that is only true for
vcpkg installations.

Signed-off-by: Silvio Traversaro <silvio.traversaro@iit.it>

Co-authored-by: Jose Luis Rivero <jrivero@osrfoundation.org>
Co-authored-by: Louise Poubel <louise@openrobotics.org>