Fix pkg_config_entry when version number is not specified
When `gz_find_package` is called without a VERSION argument, the
`<package>_FIND_VERSION` variable is empty. But in most of our
`Find*.cmake` modules, we add a pkg-config entry that assumes the
version is there.

Example:
```
gz_pkg_config_entry(GzProtobuf "protobuf >= ${GzProtobuf_FIND_VERSION}")
```

The pkg-config entry will then be "protobuf >= ", which is invalid and
causes the following entry to be treated as a version number.

The fix here removes the operator if the version is empty.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>