Use visibility hidden by default
The commit changes the gz-cmake projects to use C/C++ visibility hidden
by default. This is a potential breaking changed for projects using
gz-cmake but the benefits in terms of creating portable code and
time spend by the loader could be huge.

To avoid this change, the EXPOSE_SYMBOLS_BY_DEFAULT flag can be used
in the gz_configure_project call.

The change deprecates the HIDDEN_SYMBOLS_BY_DEFAULT flag.

Signed-off-by: Jose Luis Rivero <jrivero@osrfoundation.org>