Preserve executable permissions when installing scripts
These scripts use a "shebang" to indicate what interpreter to use when
directly invoking them from a shell, however when they are installed,
the executable permissions are dropped meaning that they actually can't
be invoked directly from a shell.

An alternative to this change might be to drop the "shebang" from the
scripts if they are intended to be invoked explicitly using the desired
interpreter, i.e. `sh tools/doc_check.sh`.

Signed-off-by: Scott K Logan <logans@cottsay.net>